Refactor code into package
