+++ src/main/java/harvard/Command.java
@@ -1,3 +1,4 @@
+package harvard;
 public abstract class Command {
     /**
      * Executes the command.
