+++ src/main/java/harvard/DukeException.java
@@ -1,3 +1,4 @@
+package harvard;
 public class DukeException extends Exception{
     public DukeException(String message) {
         super(message);
