+++ src/main/java/harvard/Event.java
@@ -1,3 +1,4 @@
+package harvard;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
