+++ src/main/java/harvard/ExitCommand.java
@@ -1,5 +1,5 @@
-public class UnmarkCommand extends Command{
-    private int index;
+package harvard;
+public class ExitCommand extends Command {
     /**
      * Returns true if the command is an exit command.
      *
@@ -7,10 +7,7 @@ public class UnmarkCommand extends Command{
      */
     @Override
     public boolean isExit() {
-        return false;
-    }
-    public UnmarkCommand(int index) {
-        this.index = index;
+        return true;
     }
 
     /**
@@ -23,12 +20,6 @@ public class UnmarkCommand extends Command{
      */
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
-        if (index >= tasks.size() || index < 0) {
-            throw new DukeException("OOPS!!! The task number does not exist.");
-        }
-        Task task = tasks.get(index);
-        task.markAsUndone();
-        ui.showUndone(task);
-        storage.save(tasks);
+        ui.displayBye();
     }
 }
