+++ src/main/java/harvard/ListCommand.java
@@ -1,3 +1,4 @@
+package harvard;
 public class ListCommand extends Command{
     /**
      * Returns true if the command is an exit command.
