+++ src/main/java/harvard/Parser.java
@@ -1,3 +1,4 @@
+package harvard;
 public class Parser {
     public static Command parse(String command) throws DukeException {
         if (command.equals("bye")) {
